More refactoring + bug fixes
