+++ src/gui/wizards/setupdevicedialog.cpp
