+++ src/gui/wizards/setupdevicedialog.h
@@ -6,8 +6,8 @@
 #include <QListWidgetItem>
 #include <QMovie>
 
-#include "../app/devicemanager.h"
-#include "../app/mqttservermanager.h"
+#include "app/devicemanager.h"
+#include "app/mqttservermanager.h"
 
 namespace Ui {
 class SetupDeviceDialog;
